Created SIOF-workaround for JR2T and T2JR links #22
