+++ quilt/core/links.cpp
@@ -1,5 +1,5 @@
 #include "include/oscillators.hpp"
-#include "include/multiscale.hpp"
+// #include "include/multiscale.hpp"
 #include "include/links.hpp"
 
 /******************************************* LINK BASE ****************************************/
@@ -14,8 +14,8 @@ LinkFactory::LinkFactory()
     add_linker(std::make_pair("base", "base"), link_maker<Link>);
     add_linker(std::make_pair("jansen-rit", "jansen-rit"), link_maker<JRJRLink>);
     
-    add_linker(std::make_pair("jansen-rit", "transducer"), link_maker<JR2TLink>);
-    add_linker(std::make_pair("transducer", "jansen-rit"), link_maker<T2JRLink>);
+    // add_linker(std::make_pair("jansen-rit", "transducer"), link_maker<JR2TLink>);
+    // add_linker(std::make_pair("transducer", "jansen-rit"), link_maker<T2JRLink>);
     
     add_linker(std::make_pair("leon-jansen-rit", "leon-jansen-rit"), link_maker<LJRLJRLink>);
 }
@@ -45,38 +45,3 @@ double LJRLJRLink::get(int axis, double now){
     return result;
 }
 
-/******************************************** MULTISCALE LINK MODELS *******************************************8*/
-
-
-double T2JRLink::get(int axis, double now){
-    // This function is called by Oscillator objects linked to this transducer
-    // during their evolution function
-    get_global_logger().log(DEBUG, "T2JRLink: getting t=" + to_string(now-delay) );
-
-    // Returns the activity of the spiking population back in the past
-    // Note that the average on the large time scale is done by Transducer::get_past()
-    double result = weight * std::static_pointer_cast<Transducer>(source)->get_past(axis, now - delay); //axis is useless
-    return result;
-}
-
-double JR2TLink::get(int axis, double now){
-
-    if (axis != 0) throw runtime_error("Jansen-Rit model can only ask for axis 0 (pyramidal neurons)");
-
-    // Returns the rate of the oscillator back in the past 
-    double v0 =  source->get_past(axis, now - delay);
-    double rate = std::static_pointer_cast<jansen_rit_oscillator>(source)->sigm(v0);
-    double result = weight * rate;
-
-    // cout << "Getting past from JRJR link" << endl;
-    // cout << "JRJR got "<<result<< endl;
-
-    //NOTE: Jansen-Rit Model is in ms^-1. Result must be converted.
-    result *= 1e3;
-
-    std::stringstream ss;
-    ss << "JR2TLink:" << "now is t=" << now << " and getting " << now-delay << ":\n v0 = " << v0 << " mV, rate = " << rate << " ms^-1, weight = " << weight << " (returning " << result << " Hz)\n";
-    get_global_logger().log(DEBUG, ss.str());
-
-    return result;
-}
